--- Assets/VRM10/vrmlib/Runtime/IndexExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ad2e7d02ba5894344a2f58d9d2157f98
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
