--- Assets/VRM10/vrmlib/Runtime/MeshExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d56612ad117ef64fb1e374d054f449b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
