+++ Assets/VRM10/vrmlib/Runtime/MeshGroup.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f3f90205124143446bab06f56827a6ee
+guid: 8d578a2f92d3c57439dc109ec5986c13
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
