--- Assets/VRM10/vrmlib/Runtime/MeshGroupExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 66fe20a63e11a3b4fb734a57c5a71839
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
