--- Assets/VRM10/vrmlib/Runtime/ModelExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: abaf884e0faa5d040ad4eae16120cb46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
