--- Assets/VRM10/vrmlib/Runtime/ModelExtensionsForCoordinates.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 382bf6be84d3135438af6a559a599fa3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
