--- Assets/VRM10/vrmlib/Runtime/ModelModifier.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a5b1c7c663b54b49913c6ade2a342c2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
