--- Assets/VRM10/vrmlib/Runtime/ModelModifierExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 54fd566dc84b51643886ce8858a671db
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
