--- Assets/VRM10/vrmlib/Runtime/VertexBufferExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c95aaad1a35e0c45a6f37f4a5604458
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
